docs: update usage and add scripts
