+++ src/routes/authRoutes.ts
@@ -1,0 +1,9 @@
+import { Router } from 'express';
+import { register, login } from '../controllers/AuthController';
+
+const router = Router();
+
+router.post('/register', register);
+router.post('/login', login);
+
+export default router;
